refactor: move invertible transformer to own file
