+++ src/safeds/data/tabular/transformation/_experimental_standard_scaler.py
@@ -2,13 +2,14 @@ from __future__ import annotations
 
 from typing import TYPE_CHECKING
 
-from safeds.data.tabular.containers import ExperimentalTable, Table
 from safeds.exceptions import NonNumericColumnError, TransformerNotFittedError, UnknownColumnNameError
 
-from ._experimental_table_transformer import ExperimentalInvertibleTableTransformer
+from ._experimental_invertible_table_transformer import ExperimentalInvertibleTableTransformer
 
 if TYPE_CHECKING:
     from sklearn.preprocessing import StandardScaler as sk_StandardScaler
+
+    from safeds.data.tabular.containers import ExperimentalTable
 
 
 class ExperimentalStandardScaler(ExperimentalInvertibleTableTransformer):
